+++ lib/ua_inspector/short_code_map.ex
@@ -15,16 +15,23 @@ defmodule UAInspector.ShortCodeMap do
       defp read_database do
         {local, _} = source()
         map = Path.join(Config.database_path(), local)
-        contents = YAML.read_file(map)
 
-        case contents do
-          {:ok, entries} ->
-            Enum.map(entries, &to_ets/1)
+        map
+        |> YAML.read_file()
+        |> parse_yaml_entries(map)
+      end
 
-          {:error, error} ->
-            _ = Logger.info("Failed to load short code map #{map}: #{inspect(error)}")
-            []
-        end
+      defp parse_yaml_entries({:ok, entries}, _) do
+        Enum.map(entries, &to_ets/1)
+      end
+
+      defp parse_yaml_entries({:error, error}, map) do
+        _ =
+          unless Config.get(:startup_silent) do
+            Logger.info("Failed to load short code map #{map}: #{inspect(error)}")
+          end
+
+        []
       end
     end
   end